apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyswisseph>=2.10.3.2",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "timezonefinder>=8.2.1",
    "uvicorn[standard]>=0.41.0",
]
